+++ conda_forge_tick/all_feedstocks.py
@@ -5,7 +5,8 @@ import github
 import logging
 
 from conda_forge_tick import sensitive_env
-from .utils import setup_logger, load, dump
+from .utils import setup_logger
+from .lazy_json_backend import load, dump
 
 logger = logging.getLogger("conda_forge_tick.all_feedstocks")
